Removed assign() and clear().
Objects are supposed to be removed from the container using Handlers, so assign() and clear() created more problems than they solved. They made handlers invalid, so if objects of this container are handled in different places they would have to be notified, thus defeating the purpose of this container. The only Container member function that is allowed to make pointers invalid is the destructor (obviously).