Addressed #1
The private Container::remove() function is in charge of destroying an element and replacing it with the last one in the container, in order not to leave holes behind.
Removed Data struct: now values and pointers to iterators are stored in two different places.
Removed Iterator since it would be the exact same thing as a pointer.
Now Handler holds a pointer to the Container to which its value belongs: Container::remove() has to be called when removing/destructing Iterators